(DE-1442) feat(TT dummy): Add svd as dummy model (#3987)
* feat: Add generate_svd_baseline and move baseline predictions / plots to utils

* refactor: Remove second for loop

* feat: Use num_recommendations

* feat: Add svd computation

* refactor: Refacto evaluate

* feat: Add svd plots
